style: Update global CSS imports to use the styles directory
- Move global CSS imports from 'globals.css' to 'styles/globals.css'
- Ensure consistency and better organization of styles.
