+++ app/page.tsx
@@ -11,7 +11,7 @@ export default function Landing() {
 				<div className="hero-overlay backdrop-blur-sm"></div>
 				<div className="hero-content text-neutral-content text-center">
 					<div className="max-w-md">
-						<img src="./aidlink.svg" alt="HirayaLink" />
+						<img src="./aidlink.svg" alt="HirayaLink" className="max-w-md" />
 						<h1 className="mb-5 text-5xl font-bold">
 							<span className="text-primary">HIRAYA</span>
 							<span>LINK</span>
